refactor: package structures
