--- app/src/main/java/pe/edu/upc/attentionapp/api/dto/request/LoginRequest.kt
@@ -1,3 +1,0 @@
-package pe.edu.upc.attentionapp.api.dto.request
-
-data class LoginRequest(var email: String, var password: String)
