+++ app/src/main/java/pe/edu/upc/attentionapp/models/Users.kt
@@ -1,7 +1,7 @@
-package pe.edu.upc.attentionapp.models.entities
+package pe.edu.upc.attentionapp.models
 
 import com.google.gson.annotations.SerializedName
-import pe.edu.upc.attentionapp.models.entities.Customer
+import pe.edu.upc.attentionapp.models.Customer
 
 class Users {
