+++ app/src/main/java/pe/edu/upc/attentionapp/network/api/AuthenticationAPI.kt
@@ -1,0 +1,10 @@
+package pe.edu.upc.attentionapp.network.api
+
+import pe.edu.upc.attentionapp.network.responses.LoginResponse
+import retrofit2.Call
+import retrofit2.http.Body
+import retrofit2.http.POST
+
+interface AuthenticationAPI {
+
+}
