+++ app/src/main/java/pe/edu/upc/attentionapp/network/responses/LoginResponse.kt
@@ -1,3 +1,3 @@
-package pe.edu.upc.attentionapp.api.dto.response
+package pe.edu.upc.attentionapp.network.responses
 
 data class LoginResponse (var message: String, var token: String)
