+++ app/src/main/java/pe/edu/upc/attentionapp/network/responses/common/DataResponse.kt
@@ -1,14 +1,15 @@
-package pe.edu.upc.attentionapp.api.dto.response.common
+package pe.edu.upc.attentionapp.network.responses.common
 
 import com.google.gson.annotations.SerializedName
 
-class CollectionResponse<T>{
+class DataResponse <T> {
+
     @SerializedName("success")
     var success: Boolean = false
 
     @SerializedName("numRows")
     var numRows: Int = 0
 
-    @SerializedName("rows")
-    lateinit var rows: ArrayList<T>
+    @SerializedName("data")
+    var data: T? = null
 }
