+++ app/src/main/java/pe/edu/upc/attentionapp/util/AttentionAppConfig.kt
@@ -1,6 +1,6 @@
 package pe.edu.upc.attentionapp.util
 
-class ConfigConstants {
+class AttentionAppConfig {
     companion object{
         const val API_V1_BASE_URL = "https://attention-api.herokuapp.com/api/v1/"
     }
