+++ app/src/main/res/layout/activity_create_account.xml
@@ -5,7 +5,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.activities.CreateAccountActivity"
+        tools:context=".controllers.activities.CreateAccountActivity"
         >
 
         <androidx.constraintlayout.widget.ConstraintLayout
